Don't schedule disposables on a different thread.
This makes it impossible to run them in a specific context. Rather assert that they
are running on the correct thread. If not fix the code up later. The problem
is that we wish to group disposables, switch to the back buffer, dispose all,
then switch back to the front. If we separately schedule each disposable
there is no garuntee they will be executed when the back buffer is active.